filtering api finished
